BatParserCo.suspend: added a function to suspend parsing (to be resumed later)
This change enables writing parsers that can give intermediate answers back
to the client and then let the client resume parsing. For example, if the data
is newline-separated and the newline detection is part of the parsing code,
each row of data can be returned immediately instead of building up a big answer
to be returned in the end.