Enforce consistency of sequence indexing in the documentation comments of .mli files.
Some modules used [a1..aN] to enumerate the elements of a linear
structure, while other used [a0..aN].
For consistency, all comments are now changed to use the [a0..aN]
convention (with the exception of BigArray, see below). I choosed
[a0..aN] over [a1..aN] to simplify description of the init, get, set
and mapi functions.

The Bigarray exception: there is a specific use case for enumerated
elements in the Bigarray modules, where the elements of an array
denote the dimensions of a multi-dimensional Bigarray structure.
For example:
  external get: ('a, 'b, 'c) t -> int array -> 'a = "caml_ba_get_generic"
    (** Read an element of a generic big array.
        [Genarray.get a [|i1; ...; iN|]] returns the element of [a]
        whose coordinates are [i1] in the first dimension, [i2] in
        the second dimension, ..., [iN] in the [N]-th dimension.

        If [a] has C layout, the coordinates must be greater or equal than 0
        and strictly less than the corresponding dimensions of [a].
        If [a] has Fortran layout, the coordinates must be greater or equal
        than 1 and less or equal than the corresponding dimensions of [a].
        Raise [Invalid_arg] if the array [a] does not have exactly [N]
        dimensions, or if the coordinates are outside the array bounds.

        If [N > 3], alternate syntax is provided: you can write
        [a.{i1, i2, ..., iN}] instead of [Genarray.get a [|i1; ...; iN|]].
        (The syntax [a.{...}] with one, two or three coordinates is
        reserved for accessing one-, two- and three-dimensional arrays
        as described below.) *)

In that case, I believe that the a1..aN indexing is really important
for comprehension.  In particular, the documentation implicitly relies
on the fact that the metavariable N is *exactly* the length of the
array (number of dimensions) : "If [N > 3]...".

Note that the "[N]-th" expression in that sentence corresponds to the
actual "-th" usage in the english language (the ordinal starting from
1), and not "the element with index N in the array". Apart from that
case which is unambiguous, the use of "-th" should probably be
avoided.

Remark: some comments used the "a1..N" syntax to refer to the elements
a1, a2, .., aN.  Here "1..N" range over the indices rather than the
elements themselves. It is slightly ambiguous, and I have rewritten
those to "a1..aN", and even "a0,a1..aN" in general.

Remark : some comments use indices for different input parameters in
fixed number (generally 2), for example [(e1 && e2)]. Those were left
unchanged.