Exhaustive testing of functions in intersection of functorized Map and polymorphic PMap interfaces.
Exhaustive testing is a bit tedious, but it already revealed some
issues in evaluation order in (P)Map.map(i), and should hopefully
provide efficient regression testing for future changes.

I hope that the feature set of Map and Pmap will eventually converge,
so that all functions are tested in both implementations.