BatInt.compare changed to external runtime primitive "caml_int_compare"
caml_int_compare is the specialized primitive for int comparison,
implemented in the ocaml runtime (byterun/ints.c). Testing reveals
that it is significantly faster (around twice faster) than using the
polymorphic comparison function.

I'm not sure however what are the implications of using an external
runtime primitive, as I'm not familiar with such low-level
considerations.