Use of a camlp4-macro INCLUDE directive to reduce .mli duplication in bat{Set,Map}.mli
Currently, BatPMap.mli is entirely included in BatMap.mli (in addition
to the functorized Map signature). I'm not sure why that choice was
made, but it's the way it is. This important code duplication led to
several misbehaviors were a function was available through BatMap.foo,
but not BatPMap.foo.

Here, an INCLUDE directive from the Camlp4MacroParser (proposing
feature similar to the C preprocessor : INCLUDE, DEFINE...) is used to
directly include batPMap.mli into batMap.mli, instead of duplicating
the code.
There is a bit of ocamlbuild magic to make it work, as ocamlbuild has
to be told to import batPMap.mli into _build/src before preprocessing
batMap.mli.