BatString.find_from: don't raise undocumented exceptions
BatString.find_from raised Invalid_argument "Empty string to search
in" when it was called with an empty string, which was undocumented
behavior and clashes what one would expect from for example List.find
((=) 0) [].

This patch changes the function to raise Not_found in that case, as it
is impossible to find a any non-empty string in an empty string. Empty
search string is handled already separately (by always finding the
substring at the offset).

This patch includes a test suite for BatString.exists,
which was the function that triggered this issue for me. It
makes use of BatString.find_from.

Signed-off-by: Erkki Seppälä <flux@modeemi.fi>