BatSplay: inconsistent behaviour of 'print' changed
BatMap.print assumes that the delimiter between the keys and values is
":". BatSplay used "," to produce an output that would be valid OCaml
code for a list. I changed the 'print' function to use ":" so that it
behaves exactly like the BatMap.print function, but added
a BatSplay.print_as_list function with the previous behavior (but no
customizable first, last and delimitors: it only prints as a list).

This issue highlights a larger problem: the 'print' function of
BatMap, and more generally of all associative collections, is not
customizable enough, as the key-value separator is imposed. I think
the best solution would be to have a more flexible print function that
would accept, instead of two separate printers for keys and values,
a single printer taking both parameters. This might come in a later
commit, but there would be compatibility issues to be dealt with.