[fix] pmap binary operators: in case on non-compatible comparisons, revert to the slower&simpler implementation respecting the strange but public interface
While it may not look so, the interface guarantee in pmap binary
operators is very restrictive. For example, the documentation of
[union m1 m2] says "equivalent t [foldi add m1 m2]". That means that:
- the comparison function used is m2's
- in case both maps have the key, the binding used is m1's

You think it doesn't make sense? Neither do I, but that's the spec,
and doing otherwise would break compatibility (which may or may not be
acceptable in this case, I'll try to discuss it on the list). The
previous implementation used heuristic_merge which did clever things
to speed up computation even in the non-compatible case. It respected
merge interface but broke the specification of the other binary
operation. This commits revert to the old "foldi add ..."
implementations.

Note that the hot path (when the comparison functions are compatible,
which is the only sane case anyway) is still as fast thanks to the use
of "merge".