change the behaviour of PMap binary operators (union,diff,intersection,merge) so that the returned map always use the comparison function of its first parameter.
This changes departs from the "equivalent to foldi add m2 m1" comment
in union documentation, and might therefore be considered
specification-breaking (though the only case where it happens, ie. the
mixing of maps with different comparison function, is already in the
realms of the Non Specified). It results in better consistency among
the diffent binary operators.