[fix] removed the camlp4 macro to handle duplication between batP{Map,Set}.mli and bat{Map,Set}.mli
The camlp4 macro provided adequate compilation support, but didn't
mesh well with the ocamldoc output. After discussion on the list,
I fell back to the simpler method of duplicating the signatures but
keeping the comments in only one place (Bat{Map,Set}.mli).

Hopefully, the correctness tests will be able to catch any divergence
between the signatures in the future.