new performance measurements for Set.Make/PSet
Currently, BatSet.Make reuses INRIA's stdlib implementation, while
PSet, for convenience, is defined as a ('a, unit) PMap. I wished to
evaluate the performance implications of this decision, which we
inherit from Extlib. Set.Make/PSet harmonization may require
implementation changes, and we should decide wether we want to
duplicate stdlib's Set implementation and factorize Set.Make/PSet (as
we did for Map), or keep reusing stdlib's functions and have
duplication for the additional features.

The performance results confirm that using a `('a, unit) map` has
a non-neglectible performance cost: on my machine, 5% to 15% for
various operations on small (500) sets, and 7% to 19% on large (10^5)
sets.