Factorized implementation of functorized Set and PSet
The PSet implementation does not rely anymore on Map. PSets are now records containing a comparing function and a balanced tree, representing the set itself.

Code responsible for handling the tree-based implementations of sets have been factorized in the Concrete module, to avoid a lot of code duplication between the functorized implementation and the polymorphic one.

There still is a dependency on standard lib's Set module, e.g. equal and compare functions are not defined in Concrete, altought this could be easily done.