change build/make_suite so that "String.foo" is a valid test name
Some inline qtests visibly have taken the habit of prefixing the test
name with the module name (I'm not sure it's that useful as the test
engine also reports the provenance module, but it is at least for
submodules). However, the quite crude test-generation program reuse
the test name as a part of an OCaml identifier, so "String.foo" would
result in a test code that is syntactically invalid. The previous
tests self-censured themselves to write "string_foo" instead, but that
may introduce confusion when the function name itself contains "_" or
"string". I added a santization pass so that all test names produce
syntactically correct OCaml identifiers.