Hide the representation of lazy lists and improve them
In particular, the 'a BatLazyList.t type is now a strict
extension of 'a list, so that the injection from the latter
to the former is Obj.magic.

As a result of hiding the underlying lazy representation,
a new "cons" constructor called lcons is added that changes
a node thunk into a lazy list.