BatFloat.round_to_string: special-casing inf/nan values to avoid a Windows bug; fixes #191
There seems to be something weird with sprintf "%f" behavior on
special floats on Windows. To avoid further issue -- and let our
Windows users have a nice quick-and-dirty display of float values --
we hardcode "inf", "-inf" and "nan". I kept the general tests
(corresponding to the Unix sprintf specifications) in case we find
a better solution in the future, such as fixing Printf directly.

Note that the bug was spotted by ygrek's Jenkins test daemon (and
noticed by thelema); Yay to continuous testing!