Fix test_file failure
This fix closes the enumeration as part of a suffix action.  This makes further calls to .next() simply raise No_more_elements instead of calling the existing next() function.  There is still a bug in Enum.group that calls .next() even after the enumeration has raised No_more_elements, maybe this is related to its use of .force() to ensure ordering.