disable -warn-error A
For a rationale against it, look at the caml-list post "A warning
about warnings" by Damien Doligez.

We should replace this by an explicit list of warnings that we
consider errors project-wise.

The reason for the rather violent current change is that the previous
patch enabling "-w A", with "-warn-error A" enabled, suddenly broke
compilation. Sweeter fixes will come later.