Test suite for BatRandom.
This initial commit tests the cloning of random enumerations and that
the enumerations properly depend on the default or used provided states.

The current implementation of BatRandom fails these tests either because
it is constructing its own seeds instead of relying on the global state
or because the clone method does not take into account that the user may
reuse the seed elsewhere in the program.