Make enumerations depend on the default PRNG state and use default clone
This commit replaces the implementation of random enumerations in
BatRandom such that this comment in BatRandom.State becomes true:

  These functions are the same as the basic functions, except that they
  use (and update) the given PRNG state instead of the default one.

Currently the enumerations in BatRandom outside of BatRandom.State
construct their own PRNG state behind the back of the user. This is
problematic since the enumerations become unreproducible and since
different enumerations by accident may be constructed from identical
or near-identical states.

Because the state is now shared, cloning can no longer be implemented by
simply copying the state. Instead clone relies on the default
implementation of BatEnum which happens to work because it lazily
buffers the stream. The change of clone is applied to BatRandom.State
too, because its enumerations also depend on external states.