Add Random.Incubator.Private_state_enums
This is an alternate implementation of BatRandom's random enums that
keeps all the previous properties of random enums (cheap clone, raises
Infinite_enum) at the cost of breaking open the random state
abstraction so that the current state of the global generator can be
"bumped" when using it to generate an enum.

The way this bump is done is to take the current state of the RNG and
use it as the input of State.make, which hashes the input to generate a
PRNG state.  This is both deterministic and repeatable without obvious
problem.

Private_state_enums.State is the same as the previous implementation
except that a copy of the input state is made before using it, so that
the RNG state is private to this enum.

Tests have been added that correspond to the existing RNG tests, which
pass trivially.