Update BatBounded to support open and closed bounds; Add interface
The bounding_of_ord_chain function could use a better name.  It passes
out-of-range values to the given functions rather than filling them in
with a single low or high default value.

bounding_of_ord could be implemented on top of bounding_of_ord_chain.  It
would simplify the code a bit but could slow the code down when large
numbers of out of range values are being created.