Merge Unix into BatUnix.
**WARNING**
For some reason the test in batteries_compattest.ml does not pass any more.
The compiler complaints that both "type error" from Legacy.Unix and the
one in the new BatUnix are not the same, although they really are.
I first though this was a result from the "module type of" construct
that I'm not familiar with, but failed to eproduce the problem with a
smaller example.
Please someone have a look!
**WARNING**