Don't copy array for BatArray.backwards enum clone
This will make it faster and enable value sharing.
Without sharing a change to the original array would be visible in the
original backwards enum, but not in its clone. This certainly is not the
expected behaviour.

See commit 8944b0d30b5d062e2f01f5bd9e73e5a9718c5a51