add an Array1.of_enum function to Bigarray, strengthen Array1.enum specification
In an ideal world we should also provide enum / of_enum pair for
higher-dimensional cases, but in practice most people only use Array1
anyway (to talk with C bindings). Note that enum already exists in the
generic-dimension case GenArray, but the new implementation is better
specified as it guarantees the expected sequential ordering.