add a .mailmap entry
I'm looking at the author list for an OUPS talk, and there is a lot of
duplicated people. If anyone would rather appear under a different
name, prefer another mail address, or keep a separation between
a young age's nick and their current endeavor, just let us know and
we'll change it.