BatList: a safe version of the Acc module, as a proof of soundness
I checked that using SafeAcc insteaf of Acc makes the unit-test
pass. If we assume this means they are observationally equivalent
(which is a long shot given the poor test coverage in batList.ml),
this proves soundness of the unsafe implementation -- note that some
mutability-informed compiler optimizations may still break this.