disable Queue.filter for the release
I have doubts about the Queue.filter interface, which is imperative:
  filter : ('a -> bool) -> 'a t -> unit

While that makes sense as a mutable operation on the Queue datatype,
the fact is that all [filter] functions (including for mutable
data-structures such as Dllist and Dynarray) return a filtered copy of
their datatype instead of mutating in place. This convention is used
by the pa_comprehension syntax extension that will happily generate
calls to filter for any data-structure module it is given, expecting
this interface.

Maybe we should consider having
  filter : ('a -> bool) -> 'a t -> 'a t
  filter_in_place : ('a -> bool) -> 'a t -> unit