harden the implementation of BatSplay against immutable-data compiler optims
BatSplay uses a bit of type-checking magic to respect the Map
interface (which requires map types to be covariant) while internally
keeping a mutable reference for (inobservable) rebalancing
mutations. The previous implementation is safe for the current
versions of the OCaml compiler, but has the fundamental issue of
mutating (through Obj magic) constructed values of a known-immutable
type.

The new implementation creates values at an immutable type, and then
magically coerce them into the covariant immutable type. This ensures
that the compiler knows (at value-creation time) that the allocated
memory is *not* immutable -- this is the same technique we use for
destination-passing-style lists.

This trick of "covariant mutable references" can be isolated as
a small submodule, which is then used (internally) throughout the
BatSplay implementation.

Thanks to Pierre Chambart for the discussion that led to this change.