Seq.Traverse: add tests
Note that the behavior of Traverse functors on lazy sequences is still
not satisfying, because they tend to force the sequence (when applying
the `sequence` function for example) more than may appear necessary.

The instance of Traverse with the Option monad
  val sequence : 'a option BatSeq.t -> 'a BatSeq.option
must necessarily force the sequence to know whether there is a None somewhere,
but there is no interface reason why the Env monad
  val sequence : ('env -> 'a) BatSeq.t -> ('env -> 'a BatSeq.t)
should force the sequence (but it does).

This is not easy to fix. I think the Monad interface is not rich
enough to allow to preserve lazyness of the container (and interleave
it with the lazyness of the monad instance). Consider the map_m
implementation:

  let rec map_m f e = match e () with
    | Nil -> M.return nil
    | Cons (x, e') ->
      f x >>= fun x ->
      map_m f e' >>= fun e' ->
      M.return (cons x e')

The reason why the sequence is forced is that (map_m f e') is
evaluated before the (cons), instead of being in the cons
right-hand-side (to be computed only when subsequent elements
are requested). It is not quite possible to have the cons first,
because it needs to be inside a M.return, and once the M.return is
done it is not possible to call (map_m) recursively.

It is not clear which interface should be provided if we wanted to
allow this. Maybe something like
  val deep_bind : 'a m -> ('a -> (forall 'c. 'c m -> 'c) -> 'b m) -> 'b m
?