synchronize BatText.nsplit to match BatString.nsplit (fixes #561)
The BatText.nsplit implementation seems to suffer from a off-by-one
offset error. The present commit changes its implementation to reuse
exactly the more battle-tested implementation of BatString.split.

A consequence of this change are two small changes of behavior, that
corresponds to non-specified aspects of BatText.nsplit and arguably
fix bugs:
- nsplit now raises and (Invalid_argugment ...) exception if the
  separator passed is empty (splitting on it makes no sense)
  (the previous implementation would loop endlessly)

- on two consecutive occurences of the separate, nsplit now inserts an
  empty text in the sequence:
    nsplit "a//b/c" "/"
  is
    ["a"; ""; "b"; "c"]
  rather than
    ["a"; "b"; "c"]
  as previously.

  This ensures that for any `sep` and `str`, we have
    `join sep (split str sep)) = str`