fix the handling of optional random-generator state in BatInnerShuffle
The behavior we want for randomness-manipulating function with an
optional parameter is that:
- if a state is passed, it is used (and mutated) for random generation
- if no state is passed, the default/global random state is used
  (and mutated) for random generation

It might seem that a parameter with default value

  ?(state = Random.get_state ())

would suffice to get this behavior, but in fact Random.get_state ()
performs a copy of the default/global random state, instead of just
returning it, so the default/global state is not mutated by calls to
shuffles (they will not be truly random).

The current implementation avoids this pitfall by calling the
global/default Random.int function directly.