Fix an implementation/specification mismatch in BatVect.insert
`BatVect.insert : int -> 'a t -> 'a t` is specified so that `insert
n r u` inserts the rope `r` between the elements of index `n` and
`n+1` in `u`. Its implementation does something different, it inserts
`r` strictly before the element of index `n` in `u`.

The present commit changes the specification to match the
implementation.

First, we expect users to have tested their program instead of
trusting (or, unfortunately, even reading) the documentation, so it is
likely that the uses of BatVect.insert with the current implementation
is correct for their needs; changing it would break user program.

Second, we argue that the implemented behavior is actually far more
natural:

- the implementation (in terms of `sub` and `concat`) is nicer -- hear
  the code!

- the ranges of integer values valid for `insert` is
  `0  .. (length u - 1)`, as one would expect, instead of
  `(-1) .. (length u - 2)` as with the current specification.

- this gives a very natural invariant between `insert` and
  `BatVect.remove : (*start*)int -> (*len*)int -> 'a t -> 'a t`:
  `u` is equal to `u |> insert n r |> remove n (length r))`.

fixes #766.