added example illustrating the BatX -> Batteries.X mapping.
For an example, I think it's better to use module that's not listed in the
"Extensions to the Standard Library" section.  That section has its own
specialized message about opening Batteries.  Using a module from this
section as an illustration has a slight risk of suggesting to users
that the BatX->Batteries.X point applies only to this section.