safe-string: fix compatibility with OCaml <=4.02
The safe-string compatibility fixes rely on the Bytes package, which
we convincingly mock under OCaml<=4.02, but also on bytes-adapted
functions or just convenience functions that are useful for code
porting: Buffer.add_subbytes, String.init, etc.

Before this commit, the codebase would simply not build on OCaml<=4.02
because of this. What the commit does is to introduce a new module,
BatBytesCompat, meant to contain backward-compatibility implementation
of those functions.