Improve toplevel printability of predefined sets
Redefining the modules (instead of using those predefined
in BatSet) was causing the printers for BatSet.Int and friends
useless.

Even with this patch it is still not possible to print any set obtained
via BatSet.Make as there is no way (that I know of) to attach a printer
to the toplevel automatically.

Closes #750