Make BatSeq.Exceptionless.combine really exceptionless
Had to change the signature since it's not possible before consumption
to know if Exceptionless.combine should return Some result or None.
So now BatSeq.Exceptionless.combine returns the longest sequence possible
(stopping when one of the input sequence ends).

BatSeq.combine still raises an exception when the two inputs have
different length though. I would prefer if none of those function ever
throw an exception (as raising an exception from a function after it has
returned brings only misery), but that would be a major overhaul that
this patch has no ambition to.

(Also fixed the unit test that was not testing Exceptionless.combine)

Closes #418