BatString: Another take on nsplit changes
Recently, [nsplit] behavior wrt. empty strings was changed to return a
unique empty string instead of an empty list, to match both the
specifications and [split_on_char] behavior.

The problem with that is it could silently broke user programs.

Therefore, this patch reverts the change to [nsplit] (but adds a
deprecation warning) and introduce the new [split_on_string] with the
intended behavior.

Note: [split_on_char] has the same issue, but there is no good fix for
it as not behaving like the Stdlib seems as problematic as not behaving
like former Batteries version. [split_on_char] change is thus regarded
as a bugfix as opposed to an API change.

Ref #845