Fix Result.t and Seq.t types not compatible with those of Stdlib
Added equivalent of batSeq with Stdlib.Seq and made it work for all range
of supported compilers.

Tested on:
- v4.02.3 (before there was a result_type in Pervasives),
- v4.07.1 (before there was a Result module in Stdlib),
- v4.09.0.

Commented out the compat-test of Bigarray that does not pass on most
versions in order to be able to test the equivalence of Result and Seq.

Also add to reintroduce some prefiltering in the Makefile as I couldn't
figure out how to have ocamlbuild prefilter the batteries-compattest file.

Closes #938