Use '=' for the underline in the changelog title
Using '-' for underines in markdown is equilavent to a H2 heading. Since
all the version headings in the changelog use the prefix '##', also H2,
they appear at the same level as the title. This confuses the
dune-release program which expects the title to be a lower heading level
than the rest of the file. Underlines made of `=` are H1 headings.